requirements.txt
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
Django==2.2
django-cors-headers==2.4.1
django-ipware==2.1.0
djangorestframework==3.9.2
mysqlclient==1.4.2.post1
pytz==2019.1
requests==2.21.0
sqlparse==0.3.0
urllib3==1.24.2
